+++ src/engine/math/matrix.rs
@@ -35,8 +35,8 @@ impl MatrixHelpers for Matrix {
         let s = pos.y / len;
         // [[c, -s, 0.0], [s, c, 0.0]]
         [
-            [c as f32, -s as f32, 0.0, 0.0],
-            [s as f32,  c as f32, 0.0, 0.0],
+            [c,  -s,   0.0, 0.0],
+            [s,   c,   0.0, 0.0],
             [0.0, 0.0, 1.0, 0.0],
             [0.0, 0.0, 0.0, 1.0]
         ].into()
